+++ Card/src/main/java/com/paypal/android/card/threedsecure/ThreeDSecureRequest.kt
@@ -1,7 +1,0 @@
-package com.paypal.android.card.threedsecure
-
-data class ThreeDSecureRequest @JvmOverloads constructor(
-    val sca: SCA = SCA.SCA_WHEN_REQUIRED,
-    val returnUrl: String,
-    val cancelUrl: String
-)