+++ Card/src/test/java/com/paypal/android/card/CardAPIUnitTest.kt
@@ -1,7 +1,6 @@
 package com.paypal.android.card
 
 import com.paypal.android.card.api.CardAPI
-import com.paypal.android.card.threedsecure.ThreeDSecureRequest
 import com.paypal.android.core.API
 import com.paypal.android.core.APIRequest
 import com.paypal.android.core.HttpMethod
@@ -86,7 +85,7 @@ class CardAPIUnitTest {
     private val cardRequest = CardRequest(
         orderID,
         card,
-        ThreeDSecureRequest(returnUrl = "return_url", cancelUrl = "cancel_url")
+       "return_url"
     )
 
     private lateinit var sut: CardAPI