+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## unreleased
 * `Card`:
-  * Update `CardRequest` to require `ThreeDSecureRequest `
+  * Remove `ThreeDSecureRequest` from `CardRequest`
+  * Update `CardRequest` to pass `return_url` and an optional `sca`
 * `PayPalUI`:
   * Fix: remove loading spinner on buttons.